+++ src/Entity/PagoPersonalSaldo.php
@@ -43,7 +43,7 @@ class PagoPersonalSaldo
     private $perfil;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $monto;
 
@@ -77,12 +77,12 @@ class PagoPersonalSaldo
         return $this;
     }
 
-    public function getMonto(): ?int
+    public function getMonto(): ?float
     {
         return $this->monto;
     }
 
-    public function setMonto(int $monto): self
+    public function setMonto(float $monto): self
     {
         $this->monto = $monto;
